Binary print added

+++ holberton.h
@@ -13,6 +13,6 @@ int concat_c(char *buffer, int character, int *chars_printed);
 int concat(char *buffer, char *s2, int *chars_printed);
 int concat_i(char *buffer, int integer, int *chars_printed);
 int concat_b(char *buffer, unsigned int integer, int *chars_printed);
-int concat_l(char *buffer, unsigned long int integer, int *chars_printed);
+int concat_l(char *buffer, long double integer, int *chars_printed);
 
 #endif
